feat: Implement generic convenience methods for add, sub, xor, and explicit
As of now all convenience methods will return the type of modifiable variable corresponding to the provided parameter type. These generic methods allow for an additional class parameter to specify the modifiable variable type to create. For example, one may now create an ModifiableUnsignedInteger with Integer parameters as simply as: add(1, ModifiableUnsignedInteger.class).